use pseudo terminals instead of pipes to preserve colored output
The limit on PTYs is much more restrictive than that on pipes. It is the
intention of the ninja developers to build projects like Chromium with
10k or more threads which exceeds the default limit for pseudo
terminals.

But pipes don't allow for colors. I want to benefit form ninja's speed
not just during builds but also during development cycles. And I am not
willing to trade Clang's colored output for all of ninja's benefits.

This commit contains a hack that ensures ninja does not abort if reading
from pseudoterminal master (PTM) fails. It seems like reading from a
pseudo terminal that hasn't characters waiting returns negative rather
than zero. This may be wrong. For a complete understanding have a look
at pts(4), posix_openpt(3) and other man pages.

Without the check for a negative return value from read(2) ninja runs
fine. With the check there are errors like

  RE2C src/lexer.ccninja: fatal: read: Input/output error

or

  [6/23] CXX build/depfile_parser.oninja: fatal: read: Input/output error

Please regard this commit as a proof of concept. In its current form it
is not suited for upstream. There should be a way for the user to decide
whether to use color or have unlimited jobs. The specialization for
pseudo-terminals should be done in a subclass of Subprocess.